requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
djoser==2.2.2
pymongo==4.7.2
django-filter==23.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7